requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
docker==5.0.3
idna==3.3
PyYAML==6.0
requests==2.27.1
urllib3==1.26.9
websocket-client==1.3.2